refactor makeNotification과 base.notification 분리
